charlib.characterizer: Reorganize
