+++ charlib/characterizer/Harness.py
@@ -29,8 +29,8 @@ class PinTestBinding:
 class Harness:
     """Characterization parameters for one path through a cell.
 
-    The primary purpose of a harness is to encode mapping of pins to
-    states. It also provides some cell-related calculations."""
+    The primary purpose of a Harness is to map pins to states. At simulation time this serves as a
+    "wiring harness", telling the TestManager how to wire up the cell for a particular test."""
 
     def __init__(self, test_manager, pin_state_map: dict) -> None:
         """Create a new Harness"""
@@ -157,161 +157,10 @@ class Harness:
         internal_charge = min(abs(q_vss_dyn), abs(q_vdd_dyn)) - time_delta * avg_current
         return internal_charge * (energy_meas_high_threshold_voltage @ u_V)
 
-
-class CombinationalHarness (Harness):
-    """A CombinationalHarness captures configuration for testing a CombinationalCell."""
-
-    def __init__(self, target_cell, test_vector) -> None:
-        """Create a new CombinationalHarness."""
-        super().__init__(target_cell, test_vector)
-        # Error if we don't have a target input port
-        if not self._target_in_port:
-            raise ValueError(f'Unable to parse target input port from test vector {test_vector}')
-
-    def plot_energy(self, settings, slews, loads, cell_name):
-        """Plot energy vs slew rate vs fanout"""
-        # TODO: Consider moving this to Pin, as all the data is eventually stored there anyways
-        figure = plt.figure()
-        figure.suptitle(f'Cell {cell_name} | Arc: {self.arc_str()}')
-
-        ax = figure.add_subplot(projection='3d')
-        ax.set_proj_type('ortho')
-
-        energy_data = []
-        for slew in slews:
-            energy_row = []
-            for load in loads:
-                energy = self._calc_internal_energy(slew, load, settings.energy_meas_high_threshold_voltage())
-                energy_row.append(float(energy.convert(settings.units.energy.prefixed_unit).value))
-            energy_data.append(energy_row)
-
-        # Expand x and y vectors to 2d arrays
-        x_data = np.repeat(np.expand_dims(slews, 1), len(loads), 1)
-        y_data = np.swapaxes(np.repeat(np.expand_dims(loads, 1), len(slews), 1), 0, 1)
-
-        # Plot energy data
-        ax.plot_surface(x_data, y_data, np.asarray(energy_data), cmap='viridis', label='Energy')
-        ax.set(xlabel=f'Slew Rate [{str(settings.units.time.prefixed_unit)}]',
-               ylabel=f'Fanout [{str(settings.units.capacitance.prefixed_unit)}]',
-               zlabel=f'Energy [{str(settings.units.energy.prefixed_unit)}]',
-               title='Energy vs. Slew Rate vs. Fanout')
-
-
-class SequentialHarness (Harness):
-    def __init__(self, test_manager, pin_state_map: dict) -> None:
-        # Parse internal storage states, clock, set, and reset out of pin mapping
-        # Note that set and reset are optional, but must be provided if present
-        # on the target cell
-        self.set = None
-        self.reset = None
-        self.clock = PinTestBinding(test_manager.clock, pin_state_map[test_manager.clock.name])
-        pin_state_map[test_manager.clock.name] = 'ignore'
-        # Set up Reset
-        if test_manager.reset:
-            self.reset = PinTestBinding(test_manager.reset, pin_state_map[test_manager.reset.name])
-            pin_state_map[test_manager.reset.name] = 'ignore'
-        # Set up Set
-        if test_manager.set:
-            self.set = PinTestBinding(test_manager.set, pin_state_map[test_manager.set.name])
-            pin_state_map[test_manager.set.name] = 'ignore'
-        # TODO: handle flop internal states
-        self.flops = []
-        super().__init__(test_manager, pin_state_map)
-
-    def short_str(self):
-        harness_str = f'{self.clock.pin.name}={self.clock.state} {super().short_str()}'
-        if self.set:
-            harness_str += f' {self.set.pin.name}={self.set.state}'
-        if self.reset:
-            harness_str += f' {self.reset.pin.name}={self.reset.state}'
-        return harness_str
-
-    @property
-    def set_direction(self) -> str:
-        if not self.set:
-            return None
-        return self.set.direction
-
-    @property
-    def reset_direction(self) -> str:
-        if not self.reset:
-            return None
-        return self.reset.direction
-    
-    def invert_set_reset(self):
-        self.set.state = self.set.state[::-1] if self.set.state else None
-        self.reset.state = self.reset.state[::-1] if self.reset.state else None
-
-    @property
-    def timing_sense_constraint(self) -> str:
-        # TODO: Check that this is correct
-        return f'{self.in_direction}_constraint'
-
-    def _timing_type_with_mode(self, mode) -> str:
-        # Determine from target input and direction
-        if self.set_direction or self.reset_direction:
-            # We're targeting set or reset
-            if mode == 'recovery':
-                if self.in_direction == 'rise':
-                    return f'{mode}_rising'
-                else:
-                    return f'{mode}_falling'
-            elif mode == 'removal':
-                if self.in_direction == 'rise':
-                    return f'{mode}_falling'
-                else:
-                    return f'{mode}_rising'
-            else:
-                return None
-        elif not self.target_in_port.pin.name in [*self.flops]:
-            # We're targeting an input port
-            if mode == 'clock':
-                if self.clock.state == '0101':
-                    return 'rising_edge'
-                else:
-                    return 'falling_edge'
-            elif mode in ['hold', 'setup']:
-                if self.in_direction == 'rise':
-                    return f'{mode}_rising'
-                else:
-                    return f'{mode}_falling'
-        # If we get here, most likely the harness isn't configured correctly
-        raise ValueError(f'Unable to determine timing type for mode "{mode}"')
-
-    @property
-    def timing_type_hold(self) -> str:
-        return self._timing_type_with_mode('hold')
-
-    @property
-    def timing_type_setup(self) -> str:
-        return self._timing_type_with_mode('setup')
-    
-    @property
-    def timing_type_recovery(self) -> str:
-        return self._timing_type_with_mode('recovery')
-    
-    @property
-    def timing_type_removal(self) -> str:
-        return self._timing_type_with_mode('removal')
-    
-    @property
-    def timing_type_clock(self) -> str:
-        return self._timing_type_with_mode('clock')
-    
-    @property
-    def timing_when(self) -> str:
-        if self.in_direction == 'rise':
-            return self.target_in_port
-        else:
-            return f'!{self.target_in_port}'
-
-    def plot_energy(self, settings, slews, loads, cell_name):
-        pass
-
 # Utilities for working with Harnesses
 def filter_harnesses_by_ports(harness_list: list, in_port, out_port) -> list:
     """Finds harnesses in harness_list which target in_port and out_port"""
-    return [harness for harness in harness_list 
+    return [harness for harness in harness_list
             if harness.target_in_port.pin == in_port
             and harness.target_out_port.pin == out_port]
